// src/app/libs/pdf/types.ts
export interface DescFontRefData {
  Type: string;
  Subtype: string;
  BaseFont: string;
  CIDToGIDMap?: string;
  CIDSystemInfo: {
    Registry: string;
    Ordering: string;
    Supplement: number;
  };
  FontDescriptor: PDFReference;
  W: [0, number[]];
}

export interface FontGlyphPosition extends GlyphPosition {
  advanceWidth?: number;
}